Fix: return ResponseEntity<BoardGame>에서 Json 양방향 참조 문제 해결 & pick했을 때 likes 변경 안되던 문제 수정

+++ src/main/java/com/swyp/boardpick/service/implement/UserBoardGameService.java
@@ -12,6 +12,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Date;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -24,9 +25,15 @@ public class UserBoardGameService {
     @Transactional
     public boolean togglePick(Long userId, Long boardGameId) {
 
+        Optional<BoardGame> optionalBoardGame = boardGameRepository.findById(boardGameId);
+        if (optionalBoardGame.isEmpty())
+            return false;
+
+        BoardGame boardGame = optionalBoardGame.get();
         // 좋아요 취소
         if (userBoardGameRepository.existsByUserIdAndBoardGameId(userId, boardGameId)) {
             userBoardGameRepository.deleteByUserIdAndBoardGameId(userId, boardGameId);
+            boardGame.decreaseLikes();
             return false;
         }
         // 좋아요
@@ -34,13 +41,13 @@ public class UserBoardGameService {
             UserBoardGame userBoardGame = new UserBoardGame();
 
             User user = userRepository.findById(userId).orElseThrow(() -> new RuntimeException("User not found"));
-            BoardGame boardGame = boardGameRepository.findById(boardGameId).orElseThrow(() -> new RuntimeException("BoardGame not found"));
 
             userBoardGame.setUser(user);
             userBoardGame.setBoardGame(boardGame);
             userBoardGame.setDate(new Date());
 
             userBoardGameRepository.save(userBoardGame);
+            boardGame.increaseLikes();
             return true;
         }
     }
